requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.1
Deprecated==1.2.18
fastapi==0.115.8
flaky==3.8.1
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
objgraph==3.6.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pillow==11.1.0
pixelmatch==0.3.0
playwright==1.50.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-trace==1.0.0
python-multipart==0.0.20
requests==2.32.3
service-identity==24.2.0
sniffio==1.3.1
starlette==0.45.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zope.interface==7.2
schedule==1.2.2
